+++ SFMLGame/MapEditorScreen.cpp
@@ -11,7 +11,10 @@ void TilePalette::create() {
 	set_texture(Textures::get("tileset"));
 	vertices.setPrimitiveType(sf::Quads);
 	vertices.resize(4 * 1);
-	set_quad(&vertices[0], 0, 0, 16 * tile_width, 16 * tile_height, scroll_x, scroll_y, 16 * tile_width, 16 * tile_height);
+	set_quad(&vertices[0], 0.f, 0.f,
+		(float) (16 * tile_width), (float) (16 * tile_height),
+		scroll_x, scroll_y, 
+		(float) (16 * tile_width), (float) (16 * tile_height));
 
 	int x = get_x();
 	int y = get_y() + 16 * tile_height;
@@ -46,7 +49,10 @@ void TilePalette::draw(sf::RenderTarget &target, sf::RenderStates states) const 
 void TilePalette::scroll(int x, int y) {
 	scroll_x += 16 * x;
 	scroll_y += 16 * y;
-	set_quad(&vertices[0], 0, 0, 16 * tile_width, 16 * tile_height, scroll_x, scroll_y, 16 * tile_width, 16 * tile_height);
+	set_quad(&vertices[0], 0.f, 0.f,
+		(float) (16 * tile_width), (float) (16 * tile_height),
+		scroll_x, scroll_y, 
+		(float) (16 * tile_width), (float) (16 * tile_height));
 }
 
 inline sf::Vector2i TilePalette::get_tile_coords(int pix_x, int pix_y) {
@@ -61,7 +67,7 @@ inline void TilePalette::select_tile(int index, sf::Vector2i tile_coords) {
 	selection_rects[index].setFillColor(sf::Color::Transparent);
 	selection_rects[index].setOutlineColor(sf::Color::White);
 	selection_rects[index].setOutlineThickness(2.0f);
-	selection_rects[index].setPosition(selections[index].x * 16 + get_x(), selections[index].y * 16 + get_y());
+	selection_rects[index].setPosition(selections[index].x * 16.f + (float) get_x(), selections[index].y * 16.f + (float) get_y());
 	selection_rects[index].setSize(sf::Vector2f(16.f, 16.f));
 }
 
@@ -132,7 +138,7 @@ void MapEditorScreen::create() {
 				game->log(ss.str());
 
 				filename = text;
-				create_map(width, height);
+				create_map((int) width, (int) height);
 				remove_component(new_panel);
 				return true;
 			});
@@ -258,10 +264,10 @@ void MapEditorScreen::create() {
 
 	create_map(10, 10);
 
-	game_view.setSize(sf::Vector2f(game->get_resolution_width(), game->get_resolution_height()));
-	game_view.setCenter(sf::Vector2f(game->get_resolution_width() / 2, game->get_resolution_height() / 2));
-	gui_view.setSize(sf::Vector2f(game->get_resolution_width(), game->get_resolution_height()));
-	gui_view.setCenter(game->get_resolution_width() / 2, game->get_resolution_height() / 2);
+	game_view.setSize(sf::Vector2f((float) game->get_resolution_width(), (float) game->get_resolution_height()));
+	game_view.setCenter(sf::Vector2f((float) game->get_resolution_width() / 2.f, (float) game->get_resolution_height() / 2.f));
+	gui_view.setSize(sf::Vector2f((float) game->get_resolution_width(), (float) game->get_resolution_height()));
+	gui_view.setCenter((float) game->get_resolution_width() / 2.f, (float) game->get_resolution_height() / 2.f);
 }
 
 void MapEditorScreen::destroy() {
@@ -301,8 +307,8 @@ void MapEditorScreen::poll_events(float elapsed_time) {
 		if (selected_component == &container) {
 			if (sf::Mouse::isButtonPressed(sf::Mouse::Button::Left)) {
 				auto mouse_position = get_mouse_game_position();
-				int x = mouse_position.x;
-				int y = mouse_position.y;
+				int x = (int) mouse_position.x;
+				int y = (int) mouse_position.y;
 
 				if (map.in_bounds(x, y)) {
 					auto tile_coord = map.get_tile_coord(x, y);
@@ -313,8 +319,8 @@ void MapEditorScreen::poll_events(float elapsed_time) {
 
 					for (int i = 0; i < 2; i++) {
 						auto selection_coords = all_selection_coords[i];
-						float tex_x = selection_coords.x;
-						float tex_y = selection_coords.y;
+						float tex_x = (float) selection_coords.x;
+						float tex_y = (float) selection_coords.y;
 
 						if (map.in_tile_bounds(tile_x, tile_y)) {
 							map.set_texture_coords(i, tile_x, tile_y, tex_x, tex_y);
@@ -333,8 +339,8 @@ void MapEditorScreen::poll_events(float elapsed_time) {
 			}
 			else if (holding_screen && sf::Mouse::isButtonPressed(sf::Mouse::Button::Middle)) {
 				auto mouse_game_position = get_mouse_game_position();
-				int dif_x = holding_start_position.x - mouse_game_position.x;
-				int dif_y = holding_start_position.y - mouse_game_position.y;
+				int dif_x = (int) (holding_start_position.x - mouse_game_position.x);
+				int dif_y = (int) (holding_start_position.y - mouse_game_position.y);
 
 				game_view.move(sf::Vector2f{ (float)dif_x, (float)dif_y });
 
@@ -356,8 +362,8 @@ void MapEditorScreen::handle_event(sf::Event &event, float elapsed_time) {
 		if (selected_component == &container) {
 			if (event.mouseButton.button == sf::Mouse::Left) {
 				auto mouse_position = get_mouse_game_position();
-				int x = mouse_position.x;
-				int y = mouse_position.y;
+				int x = (int) mouse_position.x;
+				int y = (int) mouse_position.y;
 
 				if (map.in_bounds(x, y)) {
 					auto tile_coord = map.get_tile_coord(x, y);
@@ -412,7 +418,7 @@ void MapEditorScreen::handle_event(sf::Event &event, float elapsed_time) {
 			float delta = event.mouseWheelScroll.delta;
 			int mouse_position_x = event.mouseWheelScroll.x;
 			int mouse_position_y = event.mouseWheelScroll.y;
-			float zoom = 1 - (delta * 00.1); // proportion to screen size
+			float zoom = 1 - (delta * 00.1f); // proportion to screen size
 			game_view.zoom(zoom);
 			window->setView(game_view);
 		}
@@ -442,8 +448,8 @@ void MapEditorScreen::load_map(std::string filename) {
 
 void MapEditorScreen::set_highlight_obstacles(bool highlight) {
 	highlight_obstacles = highlight;
-	for (int tile_x = 0; tile_x < map.get_tile_width(); ++tile_x) {
-		for (int tile_y = 0; tile_y < map.get_tile_height(); ++tile_y) {
+	for (unsigned int tile_x = 0; tile_x < map.get_tile_width(); ++tile_x) {
+		for (unsigned int tile_y = 0; tile_y < map.get_tile_height(); ++tile_y) {
 			if (highlight) {
 				if (map.get_tile(tile_x, tile_y).obstacle)
 					map.paint_tile(tile_x, tile_y, sf::Color::Red);
